Tutorial_3_RabbitMQ_Docker/reader/requirements.txt
pika==1.1.0
flask==1.1.2
slackclient==2.9.3
slackeventsapi==2.2.1